+++ extlib/benz/include/picrin/private/object.h
@@ -17,8 +17,14 @@ KHASH_DECLARE(env, struct identifier *, symbol *)
 KHASH_DECLARE(dict, symbol *, pic_value)
 KHASH_DECLARE(weak, struct object *, pic_value)
 
-#define OBJECT_HEADER                           \
+#ifdef PIC_USE_BITMAPGC
+# define OBJECT_HEADER                           \
   unsigned char tt;
+#else
+# define OBJECT_HEADER                           \
+  unsigned char tt;                              \
+  char gc_mark;
+#endif
 
 struct object;              /* defined in gc.c */
