+++ extlib/benz/write.c
@@ -357,7 +357,7 @@ write_core(pic_state *pic, pic_value obj, xFILE *file, struct writer_control *p)
 }
 
 static void
-write(pic_state *pic, pic_value obj, xFILE *file, int mode, int op)
+write_value(pic_state *pic, pic_value obj, xFILE *file, int mode, int op)
 {
   struct writer_control p;
 
@@ -422,10 +422,10 @@ pic_vfprintf(pic_state *pic, pic_value port, const char *fmt, va_list ap)
         xfputc(pic, '\n', file);
         break;
       case 'a':
-        write(pic, va_arg(ap, pic_value), file, DISPLAY_MODE, OP_WRITE);
+        write_value(pic, va_arg(ap, pic_value), file, DISPLAY_MODE, OP_WRITE);
         break;
       case 's':
-        write(pic, va_arg(ap, pic_value), file, WRITE_MODE, OP_WRITE);
+        write_value(pic, va_arg(ap, pic_value), file, WRITE_MODE, OP_WRITE);
         break;
       }
       break;
@@ -461,7 +461,7 @@ pic_write_write(pic_state *pic)
   pic_value v, port = pic_stdout(pic);
 
   pic_get_args(pic, "o|p", &v, &port);
-  write(pic, v, pic_fileno(pic, port), WRITE_MODE, OP_WRITE);
+  write_value(pic, v, pic_fileno(pic, port), WRITE_MODE, OP_WRITE);
   return pic_undef_value(pic);
 }
 
@@ -471,7 +471,7 @@ pic_write_write_simple(pic_state *pic)
   pic_value v, port = pic_stdout(pic);
 
   pic_get_args(pic, "o|p", &v, &port);
-  write(pic, v, pic_fileno(pic, port), WRITE_MODE, OP_WRITE_SIMPLE);
+  write_value(pic, v, pic_fileno(pic, port), WRITE_MODE, OP_WRITE_SIMPLE);
   return pic_undef_value(pic);
 }
 
@@ -481,7 +481,7 @@ pic_write_write_shared(pic_state *pic)
   pic_value v, port = pic_stdout(pic);
 
   pic_get_args(pic, "o|p", &v, &port);
-  write(pic, v, pic_fileno(pic, port), WRITE_MODE, OP_WRITE_SHARED);
+  write_value(pic, v, pic_fileno(pic, port), WRITE_MODE, OP_WRITE_SHARED);
   return pic_undef_value(pic);
 }
 
@@ -491,7 +491,7 @@ pic_write_display(pic_state *pic)
   pic_value v, port = pic_stdout(pic);
 
   pic_get_args(pic, "o|p", &v, &port);
-  write(pic, v, pic_fileno(pic, port), DISPLAY_MODE, OP_WRITE);
+  write_value(pic, v, pic_fileno(pic, port), DISPLAY_MODE, OP_WRITE);
   return pic_undef_value(pic);
 }
